chore(engine): move telemetry dtos, rename telemetry GSON mapper
