+++ engine/src/main/java/org/camunda/bpm/engine/telemetry/ApplicationServer.java
@@ -14,13 +14,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.camunda.bpm.engine.telemetry.dto;
-
-import java.util.Map;
+package org.camunda.bpm.engine.telemetry;
 
 /**
  * This class represents the data structure used for collecting information
- * about the license key issued for enterprise versions of Camunda Platform.
+ * about the application server.
  *
  * This information is sent to Camunda when telemetry is enabled.
  *
@@ -28,40 +26,15 @@ import java.util.Map;
  *      "https://docs.camunda.org/manual/latest/introduction/telemetry/#collected-data">Camunda
  *      Documentation: Collected Telemetry Data</a>
  */
-public interface LicenseKeyData {
+public interface ApplicationServer {
 
   /**
-   * The name of the customer this license was issued for.
+   * The vendor of the installed application server.
    */
-  public String getCustomer();
+  public String getVendor();
 
   /**
-   * Camunda uses different license types e.g., when one license includes usage
-   * of Cawemo enterprise.
+   * The version of the installed application server.
    */
-  public String getType();
-
-  /**
-   * The expiry date of the license in the format 'YYYY-MM-DD'.
-   */
-  public String getValidUntil();
-
-  /**
-   * A flag indicating if the license is unlimited or expires.
-   */
-  public Boolean isUnlimited();
-
-  /**
-   * A collection of features that are enabled through this license. Features
-   * could be Camunda BPM, Optimize or Cawemo.
-   */
-  public Map<String, String> getFeatures();
-
-  /**
-   * The raw license data. This combines all data fields also included in this
-   * class in the form which is stored in the license key String. Note, that
-   * this is not the license key as issued to the customer but only contains the
-   * plain-text part of it and not the encrypted key.
-   */
-  public String getRaw();
+  public String getVersion();
 }
