+++ engine/src/main/java/org/camunda/bpm/engine/telemetry/Command.java
@@ -14,38 +14,31 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.camunda.bpm.engine.telemetry.dto;
+package org.camunda.bpm.engine.telemetry;
+
+import org.camunda.bpm.engine.ManagementService;
 
 /**
- * This class represents the data structure of information about the Camunda
- * product.
+ * This class represents the data structure used for collecting information
+ * about executed commands for telemetry data. A command is an action usually
+ * triggered by a user and performed by the engine. This class counts the number
+ * of executions per command.
  *
  * This information is sent to Camunda when telemetry is enabled.
+ *
+ * When used for telemetry data collection, all command execution counts reset
+ * on sending the data. Retrieval through
+ * {@link ManagementService#getTelemetryData()} will not reset the counter.
  *
  * @see <a href=
  *      "https://docs.camunda.org/manual/latest/introduction/telemetry/#collected-data">Camunda
  *      Documentation: Collected Telemetry Data</a>
  */
-public interface Product {
+public interface Command {
 
   /**
-   * The name of the product (i.e., Camunda BPM Runtime).
+   * The count of this command i.e., how often did the engine engine execute
+   * this command.
    */
-  public String getName();
-
-  /**
-   * The version of the process engine (i.e., 7.X.Y).
-   */
-  public String getVersion();
-
-  /**
-   * The edition of the product (i.e., either community or enterprise).
-   */
-  public String getEdition();
-
-  /**
-   * Information about the technical internals and the environment of the
-   * Camunda installation.
-   */
-  public Internals getInternals();
+  public long getCount();
 }
