+++ engine/src/main/java/org/camunda/bpm/engine/telemetry/Internals.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.camunda.bpm.engine.telemetry.dto;
+package org.camunda.bpm.engine.telemetry;
 
 import java.util.Map;
 import java.util.Set;
